refactor: mv shared code from storefront and rm storefront
