+++ web-marketplace/src/components/organisms/SellOrdersTable/SellOrdersTable.Row.tsx
@@ -15,7 +15,6 @@ import DenomIcon from 'components/molecules/DenomIcon';
 import DenomLabel from 'components/molecules/DenomLabel';
 
 import { getAccountUrl } from '../../../lib/block-explorer';
-import { NormalizedSellOrder } from '../../../pages/Marketplace/Storefront/Storefront.types';
 import { Link } from '../../atoms';
 import WithLoader from '../../atoms/WithLoader';
 import {
