+++ web-marketplace/src/features/marketplace/CancelSellOrderFlow/CancelSellOrderFlow.constants.ts
@@ -1,0 +1,6 @@
+import { msg } from '@lingui/macro';
+
+export const CANCEL_SELL_ORDER_HEADER = msg`Your sell order was cancelled`;
+export const CANCEL_SELL_ORDER_TITLE = msg`Sell Order #`;
+export const CANCEL_SELL_ORDER_ACTION = msg`Cancel`;
+export const CANCEL_SELL_ORDER_BUTTON = msg`View all sell orders`;
