--- web-marketplace/src/pages/Marketplace/Storefront/Storefront.constants.ts
@@ -1,11 +1,0 @@
-import { msg } from '@lingui/macro';
-
-export const BUY_SELL_ORDER_TITLE = msg`Buy`;
-export const BUY_SELL_ORDER_ACTION = msg`BUY`;
-
-export const CANCEL_SELL_ORDER_HEADER = msg`Your sell order was cancelled`;
-export const CANCEL_SELL_ORDER_TITLE = msg`Sell Order #`;
-export const CANCEL_SELL_ORDER_ACTION = msg`Cancel`;
-export const CANCEL_SELL_ORDER_BUTTON = msg`View all sell orders`;
-
-export const STOREFRONT_TWITTER_TEXT = msg`I just purchased carbon credits on #RegenMarketplace with @regen_network. Let's work together to reduce our carbon footprint. Join me in investing in a regenerative future! `;
