--- web-marketplace/src/pages/Marketplace/Storefront/index.ts
@@ -1,1 +1,0 @@
-export { Storefront as default } from './Storefront';
